refactor[#6] : SnippetOutput -> SnippetDisplay 컴포넌트로 리팩토링
- htmlContent 생성하던 비즈니스 로직 useSnippetContent 커스텀훅으로 분리
