+++ app/src/features/snippet/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./useCopySnippetImage";
+export * from "./useSnippetContent";
