+++ app/src/features/snippet/ui/SnippetDisplay.tsx
@@ -1,76 +1,10 @@
 "use client";
 
-import styles from "./snippet.module.css";
-import { useAppSelector } from "@/hooks/redux-hooks";
-import { useEffect, useState } from "react";
-import { markdownProcessor } from "../lib/markdownProcessor";
+import styles from "./styles.module.css";
+import { useSnippetContent } from "../lib";
 
 export const SnippetDisplay = () => {
-  const snippetInput = useAppSelector((state) => state.snippet.text);
-  const showLineNumbers = useAppSelector(
-    (state) => state.setting.showLineNumbers
-  );
-  const [backgroundColor, setBackgroundColor] = useState<string>("inherit");
-
-  const snippetSetting = useAppSelector((state) => state.setting);
-  const [htmlContent, setHtmlContent] = useState<string>("");
-
-  useEffect(() => {
-    (async function () {
-      const markdown = markdownProcessor.preprocessMarkdown(
-        snippetInput,
-        snippetSetting
-      );
-      const result = await markdownProcessor.processMarkdown(
-        markdown,
-        snippetSetting.theme
-      );
-      setHtmlContent(result);
-    })();
-  }, [snippetInput, snippetSetting]);
-
-  useEffect(() => {
-    if (!htmlContent) return;
-
-    const $codeBlock = document.querySelector("code");
-    const $codeBlockWrapper = document.querySelector("pre");
-    if (!$codeBlock || !$codeBlockWrapper) return;
-
-    // 백그라운드 테마 색상 설정
-
-    setBackgroundColor(getComputedStyle($codeBlockWrapper).backgroundColor);
-
-    const children = Array.from($codeBlock.children);
-
-    const newChildren = children.map((child, idx) => {
-      const $lineWrapper = document.createElement("div");
-      $lineWrapper.className = "lineWrapper";
-
-      const $lineNumber = document.createElement("div");
-      const lineNumber = Number(showLineNumbers || 1) + idx;
-      $lineNumber.textContent = String(lineNumber);
-      $lineNumber.className = "lineNumber";
-
-      const codeSpan = child.matches("span[data-line]")
-        ? child
-        : (child.lastChild as HTMLSpanElement);
-
-      $lineWrapper.appendChild($lineNumber);
-      $lineWrapper.appendChild(codeSpan);
-
-      return $lineWrapper;
-    });
-
-    $codeBlock.innerHTML = "";
-    newChildren.forEach((newChild) => {
-      $codeBlock.appendChild(newChild);
-    });
-
-    const $snippetOutput = $codeBlock.parentElement
-      ?.parentElement as HTMLElement;
-
-    setHtmlContent($snippetOutput?.outerHTML);
-  }, [htmlContent, showLineNumbers]);
+  const { htmlContent, codeThemeBackgroundColor } = useSnippetContent();
 
   if (!htmlContent) {
     return (
@@ -86,7 +20,7 @@ export const SnippetDisplay = () => {
     <section
       className={styles.snippetOutputWrapper}
       style={{
-        backgroundColor,
+        backgroundColor: codeThemeBackgroundColor,
       }}
     >
       <div
